# swagger.yaml
swagger: "2.0"
info:
  title: "Tradee API"
  description: "API for Tradee marketplace"
  version: "1.0.0"
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
    description: "JWT Bearer token. Format: Bearer <token>"
  Cookie:
    type: "apiKey"
    name: "jwt_token"
    in: "cookie"
    description: "JWT token stored in cookie"
paths:
  /api/categories:
    get:
      summary: "Get all available categories"
      description: "Retrieve a list of all available auction categories"
      responses:
        200:
          description: "List of available auction categories"
          schema:
            type: array
            items:
              type: string
              description: "Category name"
            example: ["clothing", "furniture", "electronics"]
        500:
          description: "Server error"
  /api/auctions:
    get:
      summary: "Get auctions with optional filters"
      description: "Get auctions with optional category filter, sorting, and pagination"
      parameters:
        - name: category
          in: query
          type: string
          description: "Filter by category name (e.g., 'clothing', 'furniture')"
          required: false
        - name: query
          in: query
          type: string
          description: "Search query for semantic search across auction names"
          required: false
        - name: count
          in: query
          type: integer
          default: 6
          description: "Number of auctions to return"
        - name: offset
          in: query
          type: integer
          default: 0
          description: "Number of auctions to skip (for pagination)"
        - name: sort
          in: query
          type: string
          enum: ["published_at", "views", "price"]
          default: "published_at"
          description: "Sort by column"
        - name: owner_id
          in: query
          type: integer
          description: "Filter auctions by owner user ID"
          required: false
      responses:
        200:
          description: "List of auctions"
          schema:
            type: array
            items:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                category:
                  type: integer
                image_small:
                  type: string
                image_regular:
                  type: string
                published_at:
                  type: string
                auction_time:
                  type: integer
                views:
                  type: integer
                owner_id:
                  type: integer
                location:
                  type: string
                condition:
                  type: string
                published:
                  type: boolean

    post:
      summary: "Create a new auction"
      description: "Create a new auction with multipart form data"
      security:
        - Bearer: []
        - Cookie: []
      consumes:
        - multipart/form-data
      parameters:
        - name: name
          in: formData
          type: string
          required: true
          description: "Auction title"
        - name: description
          in: formData
          type: string
          required: true
          description: "Detailed auction description"
        - name: category
          in: formData
          type: integer
          required: true
          description: "Category ID"
        - name: price
          in: formData
          type: number
          required: true
          description: "Starting price for the auction"
        - name: auction_time
          in: formData
          type: integer
          required: true
          description: "Auction duration in seconds"
        - name: location
          in: formData
          type: string
          required: true
          description: "Item location"
        - name: condition
          in: formData
          type: string
          required: true
          description: "Item condition (e.g., 'new', 'used', 'like new')"
        - name: published
          in: formData
          type: boolean
          required: true
          description: "Whether to publish the auction immediately"
        - name: image_large
          in: formData
          type: file
          required: true
          description: "Large auction image (JPG, JPEG, PNG, or WEBP)"
      responses:
        200:
          description: "Auction created successfully"
          schema:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
        400:
          description: "Invalid request format, missing required fields, invalid image format, or no image provided"
        401:
          description: "Authentication token missing or invalid"
  /api/auctions/{auction_id}:
    get:
      summary: "Get auction details"
      description: "Get details for a specific auction"
      parameters:
        - name: auction_id
          in: path
          type: integer
          required: true
          description: "The auction ID"
      responses:
        200:
          description: "Auction details"
          schema:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              description:
                type: string
              price:
                type: number
              category:
                type: integer
              image_small:
                type: string
              image_regular:
                type: string
              published_at:
                type: string
              auction_time:
                type: integer
              views:
                type: integer
              owner_id:
                type: integer
              location:
                type: string
              condition:
                type: string
              published:
                type: boolean
        404:
          description: "Auction not found"

    put:
      summary: "Update an existing auction"
      description: "Update an existing auction with multipart form data"
      security:
        - Bearer: []
        - Cookie: []
      consumes:
        - multipart/form-data
      parameters:
        - name: auction_id
          in: path
          type: integer
          required: true
          description: "The auction ID to update"
        - name: name
          in: formData
          type: string
          description: "(Optional) Auction title"
        - name: description
          in: formData
          type: string
          description: "(Optional) Detailed auction description"
        - name: category
          in: formData
          type: integer
          description: "(Optional) Category ID"
        - name: price
          in: formData
          type: number
          description: "(Optional) Price"
        - name: auction_time
          in: formData
          type: integer
          description: "(Optional) Auction duration in seconds"
        - name: location
          in: formData
          type: string
          description: "(Optional) Item location"
        - name: condition
          in: formData
          type: string
          description: "(Optional) Item condition"
        - name: published
          in: formData
          type: boolean
          description: "(Optional) Whether auction is published"
        - name: image_large
          in: formData
          type: file
          description: "(Optional) Updated auction image (JPG, JPEG, PNG, or WEBP)"
      responses:
        200:
          description: "Auction updated successfully"
          schema:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
        400:
          description: "Invalid field types, invalid image format, or no fields provided to update"

          delete:
            summary: "Delete an auction"
            description: "Delete an auction"
            security:
              - Bearer: []
              - Cookie: []
            parameters:
              - name: auction_id
                in: path
                type: integer
                required: true
                description: "The auction ID to delete"
            responses:
              200:
                description: "Auction deleted successfully"
                schema:
                  type: object
                  properties:
                    message:
                      type: string
              400:
                description: "Invalid auction ID or other errors"
              401:
                description: "Authentication token missing or invalid"
              403:
                description: "Unauthorized to delete this auction"
              404:
                description: "Auction not found"
        401:
          description: "Authentication token missing or invalid"
        403:
          description: "Unauthorized to update this auction"
        404:
          description: "Auction not found"

    delete:
      summary: "Delete an auction"
      description: "Delete an auction"
      security:
        - Bearer: []
        - Cookie: []
      parameters:
        - name: auction_id
          in: path
          type: integer
          required: true
          description: "The auction ID to delete"
      responses:
        200:
          description: "Auction deleted successfully"
          schema:
            type: object
            properties:
              message:
                type: string
        400:
          description: "Invalid auction ID or other errors"
        401:
          description: "Authentication token missing or invalid"
        403:
          description: "Unauthorized to delete this auction"
        404:
          description: "Auction not found"
  /api/auctions/{auction_id}/bids:
    get:
      summary: "Get bids for an auction"
      description: "Get all bids for a specific auction"
      parameters:
        - name: auction_id
          in: path
          type: integer
          required: true
          description: "The auction ID"
      responses:
        200:
          description: "List of bids for the auction"
          schema:
            type: array
            items:
              type: object
              properties:
                id:
                  type: integer
                price:
                  type: number
                first_name:
                  type: string
                last_name:
                  type: string
                published_at:
                  type: string
        404:
          description: "No bids found for this auction"
  /api/auctions/{auction_id}/bid:
    post:
      summary: "Place a bid on an auction"
      description: "Place a bid on an auction"
      security:
        - Bearer: []
        - Cookie: []
      parameters:
        - name: auction_id
          in: path
          type: integer
          required: true
          description: "The auction ID to bid on"
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              amount:
                type: number
                description: "The bid amount (must be higher than current highest bid)"
            required:
              - amount
      responses:
        200:
          description: "Bid placed successfully"
          schema:
            type: object
            properties:
              message:
                type: string
        400:
          description: "Invalid bid amount, auction not found, auction is not active, user is the owner of the auction, bid is too low, or invalid format"
        401:
          description: "Authentication token missing or invalid"
        404:
          description: "Auction not found"
  /api/auctions/remove_published:
    put:
      summary: "Mark an expired auction as no longer published"
      description: "Mark an expired auction as no longer published"
      security:
        - Bearer: []
        - Cookie: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              auction_id:
                type: integer
                description: "The auction ID to update"
            required:
              - auction_id
      responses:
        200:
          description: "Auction status updated successfully"
          schema:
            type: object
            properties:
              message:
                type: string
        400:
          description: "Invalid request format, auction not found, auction is unpublished, or auction is still active"
        401:
          description: "Authentication token missing or invalid"
        404:
          description: "Auction not found"
  /api/users/{user_id}/recent:
    get:
      summary: "Get recent user activity"
      description: "Get recent actions for a specific user"
      parameters:
        - name: user_id
          in: path
          type: integer
          required: true
          description: "The user ID"
      responses:
        200:
          description: "A list of recent actions for the user"
          schema:
            type: array
            items:
              type: object
              properties:
                action_type:
                  type: string
                user_id:
                  type: integer
                auction_id:
                  type: integer
                auction_name:
                  type: string
                created_auction_name:
                  type: string
                bid_price:
                  type: number
                seconds_ago:
                  type: integer
                seconds_since_created:
                  type: integer
                views:
                  type: integer
                action_time:
                  type: string
        404:
          description: "User not found or no recent actions for the user"
...